+++ CHANGES.rst
@@ -3,11 +3,14 @@ Changelog of lizard-nxt client
 
 Unreleased (4.2.0) (XXXX-XX-XX)
 -------------------------------
--
+
+- When user switches back to map context after having selected more
+  than one asset while in db context, the box.type and tool switch to
+  'multi-point'.
 
 
 Release 4.1.16 (2016-10-13)
----------------------
+---------------------------
 
 - Annotation default not present in portal, if present and active: add and turn
   on.